requirements.txt
cohere
datasets
python-dotenv